moving files out of custom_data
