+++ src/app/Main.java
@@ -1,0 +1,4 @@
+package app;
+
+public class Main {
+}
