--- src/app/custom_data/Main.java
@@ -1,4 +1,0 @@
-package app.custom_data;
-
-public class Main {
-}
